nthreads
Former-commit-id: fe84c739090377f50f7414dc9c5e6b3e0defa96c [formerly 71ea9a950e65b2123cd56dac5ab0f276dc617cd4] [formerly 5f7689f4c69ab3a048f4174e40722675d68faee5 [formerly fad07d8f9d5dd7d87111f91c66e9a84fcca5c863]] [formerly ac4faeeb6789003ad4627dbdb88ebfe89a75fd1f [formerly 7d2ca322f0bc0dcf23fd7f87adfe2de904985da3] [formerly 5fdcf537f469663d359dfb6b964341c2e942ff38 [formerly a2c83019f6b9f41d764dc79191bcc77e37513ad1]]]
Former-commit-id: 00b33a7be23bf85b6634e8bb6792462f081b36fe [formerly be800c16181a50e4d2640dc92e58d04b42150a76] [formerly 5fdcf537f469663d359dfb6b964341c2e942ff38]
Former-commit-id: 4114badeba3cd163ff3e34984090efe43bc51f39 [formerly cf2edfe6c2181a711ac6170354df5b2c72396bbf]
Former-commit-id: fef600bec667accc9217037d2681e8e18cb3a754

+++ sqanti_qc.py
@@ -549,38 +549,44 @@ def correctionPlusORFpred(args, chrom_names):
 	 	os.makedirs(gmst_dir)
 
 	subprocess.call(["perl", utilitiesPath+"gmst/gmst.pl", corrFASTA , '-faa', '--strand','direct', '--fnn', '--output',  os.path.splitext(os.path.basename(args.isoforms))[0]+"_corrected"], cwd=gmst_dir) 
-	os.rename(gmst_dir+os.path.basename(ORF), ORF) 
-	gc.collect()
+	if os.path.isfile(gmst_dir+os.path.basename(ORF)):
+		os.rename(gmst_dir+os.path.basename(ORF), ORF) 
 
-	# Modifying ORF sequences by removing sequence before ATG
+		# Modifying ORF sequences by removing sequence before ATG
 
-	ORFdicc = fasta_parser(ORF)	
-	corrORF_file = open(corrORF, "w")
+		ORFdicc = fasta_parser(ORF)	
+		corrORF_file = open(corrORF, "w")
 
-	orfLenDicc = {}
-	for ID in ORFdicc:
-		seq = ORFdicc[ID]
-		l = len(ID.split("\t"))
-		pos = seq.find('M')
-		if pos != -1:
-			corr_seq = seq[pos:]
-			cds_start = int(ID.split("\t")[l-1].split("|")[4])+pos*3
-			cds_end = int(ID.split("\t")[l-1].split("|")[5])  
-			strand = ID.split("\t")[l-1].split("|")[3]
-			ID_mod = ID.split()[0].strip()+"\t"+"isoform|GeneMark.hmm|"+ str(len(corr_seq)) + "_aa|" + strand + "|" + str(cds_start) +"|" + str(cds_end)
-			corrORF_file.write(">"+ID_mod+"\n"+corr_seq+"\n")
-			orfLenDicc[ID.split()[0].strip()] = myQueryProteins(cds_start, (cds_start+len(corr_seq)*3-1), len(corr_seq))
-	
-	corrORF_file.close()
+		orfLenDicc = {}
+		for ID in ORFdicc:
+			seq = ORFdicc[ID]
+			l = len(ID.split("\t"))
+			pos = seq.find('M')
+			if pos != -1:
+				corr_seq = seq[pos:]
+				cds_start = int(ID.split("\t")[l-1].split("|")[4])+pos*3
+				cds_end = int(ID.split("\t")[l-1].split("|")[5])  
+				strand = ID.split("\t")[l-1].split("|")[3]
+				ID_mod = ID.split()[0].strip()+"\t"+"isoform|GeneMark.hmm|"+ str(len(corr_seq)) + "_aa|" + strand + "|" + str(cds_start) +"|" + str(cds_end)
+				corrORF_file.write(">"+ID_mod+"\n"+corr_seq+"\n")
+				orfLenDicc[ID.split()[0].strip()] = myQueryProteins(cds_start, (cds_start+len(corr_seq)*3-1), len(corr_seq))
+		
+		corrORF_file.close()
+		del ORFdicc
+		os.remove(ORF)
+
+
+	else:
+		corrORF_file = open(corrORF, "w")
+		corrORF_file.close()
+		orfLenDicc={}
+
 
 	if len(orfLenDicc)==0:
  		sys.stderr.write("WARNING: All input isoforms were predicted as non-coding")
-
 	
-	os.remove(ORF)
 	os.rename(corrORF, ORF) 
 
-	del ORFdicc
 	gc.collect()
 	return(orfLenDicc)
 
@@ -1895,7 +1901,7 @@ def main():
 	if args.dir==None:
 		args.dir = os.getcwd() 
 	else:
-		if (os.path.isdir(args.dir)==False):
+		if (os.path.isdir(os.path.abspath(args.dir))==False):	
 			sys.stderr.write("ERROR: '%s' directory doesn't exist\n" %(args.dir))
 			sys.exit()
 		else:
